Add files via upload
1.	Introduction

This project is based on an online web portal for motor servicing systems for local service stations. It works as a mediator between local garages, service stations, and users of vehicles. It is a helpful platform to overcome the interactivity issues in the physical system of appearance of the user of vehicle for various issues in their vehicle.

This Web App is specially used to give the user as a customer of the vehicle and also service station owner to interact with each other and applying various functionalities and services and track their status of activities and the final invoice of thecustomer.

Especially this application is based on new platforms like ReactJS and J2EE with Database support, Due to which various functioning and activities, modules are done in our web system due to which user able to do Registration & login, Mechanic & Service provider also get access to the system using login credentials, Admin has to control all the authentications and permissions with other functioning like assign mechanic, service provider and accept the user service & inquiry request.